move testfactory to rivershared
This will be useful for tests in other River libraries (riverui,
riverpro).
