+++ go.mod
@@ -29,3 +29,5 @@ require (
 	golang.org/x/crypto v0.17.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
+
+replace github.com/riverqueue/river/rivershared => ./rivershared
