+++ rivershared/go.mod
@@ -5,6 +5,9 @@ go 1.21
 toolchain go1.23.0
 
 require (
+	github.com/riverqueue/river v0.11.4
+	github.com/riverqueue/river/riverdriver v0.11.4
+	github.com/riverqueue/river/rivertype v0.11.4
 	github.com/stretchr/testify v1.9.0
 	go.uber.org/goleak v1.3.0
 	golang.org/x/mod v0.9.0
